Add files via upload
Take a ZIP file of images and process them using a library built into python that you need to learn how to use. A ZIP file takes several different files and compresses them, thus saving space into one single file. The files in the ZIP file we provide are newspaper images. The task is to write python code which allows one to search through the images looking for the occurrences of keywords and faces. E.g. if you search for "pizza," it will return a contact sheet of all of the faces which were located on the newspaper page which mentions "pizza."  Used OpenCV to detect faces, your ability to use tesseract to do optical character recognition, and your ability to use PIL to composite images together into contact sheets.